Newton 切线迭代法完成

+++ solve_equation/constFeild.h
@@ -1,2 +1,2 @@
 #pragma once
-const float e = 2.718281828; // 对数常量
+const double e = 2.718281828; // 对数常量
